third_party/blink/web_tests/fast/events/window-errorevent-01-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: DumpErrorEventAndPreventDefault is not defined
This test should trigger 'window.onerror', and successfully handle the error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
